Sets UI to refresh automatically
The code for deleteHook and addOrUpdateHook already includes a call to await fetchHooks(); after a successful operation, which is meant to re-fetch and re-render the list. However, if the refresh isn't visually obvious, it can be confusing.

To make the refresh more explicit and provide better visual feedback, adds a line to clear the list and show a "Refreshing list..." message right before fetchHooks() is called after a successful add/update/delete operation.